refactor(lpss): 调整 lpss 头文件
1. 将 MTP 部分的内容从 rsd 头文件中剥离，独立成 rmtp
2. 将 guid 相关内容提出至 util 中
