+++ modules/lpss/include/rmvl/lpss/node.hpp
@@ -11,10 +11,7 @@
 
 #pragma once
 
-#include <chrono>
 #include <condition_variable>
-#include <shared_mutex>
-#include <thread>
 
 #include "rmvlmsg/std/time.hpp"
 
@@ -23,6 +20,7 @@
 #endif
 
 #include "details/node_rsd.hpp"
+#include "details/node_rmtp.hpp"
 
 namespace rm {
